cleanup the problems hidden by the buggy RouteResolver type
