+++ app/ui-react/syndesis/src/modules/integrations/IntegrationCreatorApp.tsx
@@ -1,4 +1,5 @@
 import { ALL_STEPS, createStep, DATA_MAPPER } from '@syndesis/api';
+import { StepKind } from '@syndesis/models';
 import { Breadcrumb } from '@syndesis/ui';
 import * as React from 'react';
 import { Route, Switch } from 'react-router';
@@ -20,7 +21,9 @@ const addStepPage = (
         ...s,
       })
     }
-    apiProviderHref={resolvers.create.configure.editStep.apiProvider.review}
+    apiProviderHref={(step, p, s) =>
+      resolvers.create.configure.editStep.apiProvider.review()
+    }
     connectionHref={(step, params, state) =>
       resolvers.create.configure.editStep.connection.configureAction({
         actionId: step.action!.id!,
@@ -32,11 +35,11 @@ const addStepPage = (
     filterHref={resolvers.create.configure.editStep.basicFilter}
     getAddMapperStepHref={(position, params, state) =>
       resolvers.create.configure.addStep.dataMapper({
-        position,
+        position: `${position}`,
         step: {
           ...createStep(),
           ...ALL_STEPS.find(s => s.stepKind === DATA_MAPPER),
-        },
+        } as StepKind,
         ...params,
         ...state,
       })
@@ -80,7 +83,7 @@ const saveIntegrationPage = (
         ...s,
       })
     }
-    postPublishHref={resolvers.integration.details}
+    postPublishHref={p => resolvers.integration.details({ ...p })}
   />
 );
