+++ app/ui-react/syndesis/src/modules/integrations/components/IntegrationEditorStepAdder.tsx
@@ -40,7 +40,10 @@ export interface IIntegrationEditorStepAdderProps {
    * @param step - the integration step object that should be edited
    */
   // tslint:disable-next-line:react-unused-props-and-state
-  configureStepHref: (stepIdx: number, step: Step) => H.LocationDescriptor;
+  configureStepHref: (
+    stepIdx: number,
+    step: Step
+  ) => H.LocationDescriptorObject;
 }
 
 /**
