+++ app/ui-react/syndesis/src/modules/integrations/resolvers.ts
@@ -1,11 +1,6 @@
 /* tslint:disable:object-literal-sort-keys no-empty-interface */
-import { getEmptyIntegration, getStep } from '@syndesis/api';
-import {
-  ConnectionOverview,
-  IIntegrationOverviewWithDraft,
-  Integration,
-  StepKind,
-} from '@syndesis/models';
+import { getEmptyIntegration } from '@syndesis/api';
+import { IIntegrationOverviewWithDraft } from '@syndesis/models';
 import {
   makeResolver,
   makeResolverNoParams,
@@ -14,54 +9,18 @@ import {
 import {
   IBaseRouteParams,
   IBaseRouteState,
-  IConfigureActionRouteParams,
-  IConfigureActionRouteState,
-  IConfigureStepRouteParams,
-  IConfigureStepRouteState,
-  IDataMapperRouteParams,
-  IDataMapperRouteState,
-  IDescribeDataShapeRouteParams,
-  IDescribeDataShapeRouteState,
+  IEditorIndex,
   ISaveIntegrationRouteParams,
   ISaveIntegrationRouteState,
-  ISelectActionRouteParams,
-  ISelectActionRouteState,
   ISelectConnectionRouteParams,
   ISelectConnectionRouteState,
-  ITemplateStepRouteParams,
-  ITemplateStepRouteState,
-  stepRoutes,
 } from './components/editor/interfaces';
+import { makeEditorResolvers } from './components/editor/makeEditorResolvers';
 import {
   IDetailsRouteParams,
   IDetailsRouteState,
 } from './pages/detail/interfaces';
 import routes from './routes';
-
-interface IEditorIndex {
-  flowId: string;
-  integration: Integration;
-}
-
-interface IEditorSelectConnection extends IEditorIndex {
-  position: string;
-}
-
-interface IEditorSelectAction extends IEditorSelectConnection {
-  connection: ConnectionOverview;
-}
-
-interface IEditorConfigureAction extends IEditorSelectAction {
-  actionId: string;
-  step?: string;
-  updatedIntegration?: Integration;
-}
-
-interface IEditorConfigureStep extends IEditorIndex {
-  position: string;
-  step: StepKind;
-  updatedIntegration?: Integration;
-}
 
 export const configureIndexMapper = ({
   flowId,
@@ -75,199 +34,6 @@ export const configureIndexMapper = ({
     integration,
   } as IBaseRouteState,
 });
-
-export const configureSelectConnectionMapper = ({
-  position,
-  ...rest
-}: IEditorSelectConnection) => {
-  const { params, state } = configureIndexMapper(rest);
-  return {
-    params: {
-      ...params,
-      position,
-    } as ISelectConnectionRouteParams,
-    state,
-  };
-};
-
-export const configureSelectActionMapper = ({
-  connection,
-  ...rest
-}: IEditorSelectAction) => {
-  const { params, state } = configureSelectConnectionMapper(rest);
-  return {
-    params: {
-      ...params,
-      connectionId: connection.id!,
-    } as ISelectActionRouteParams,
-    state: {
-      ...state,
-      connection,
-    } as ISelectActionRouteState,
-  };
-};
-
-export const configureConfigureActionMapper = ({
-  actionId,
-  flowId,
-  step,
-  integration,
-  updatedIntegration,
-  position,
-  ...rest
-}: IEditorConfigureAction) => {
-  const { params, state } = configureSelectActionMapper({
-    ...rest,
-    flowId,
-    integration,
-    position,
-  });
-  const positionAsNumber = parseInt(position, 10);
-  const stepObject = getStep(integration, flowId, positionAsNumber) || {};
-  return {
-    params: {
-      ...params,
-      actionId,
-      step: `${step || 0}`,
-    } as IConfigureActionRouteParams,
-    state: {
-      ...state,
-      updatedIntegration,
-      configuredProperties: stepObject.configuredProperties || {},
-    } as IConfigureActionRouteState,
-  };
-};
-
-export const configureDescribeDataShapeMapper = ({
-  position,
-  step,
-  updatedIntegration,
-  ...rest
-}: IEditorConfigureStep) => {
-  const { params, state } = configureIndexMapper(rest);
-  return {
-    params: {
-      ...params,
-      position,
-    } as IDescribeDataShapeRouteParams,
-    state: {
-      ...state,
-      step,
-      updatedIntegration,
-    } as IDescribeDataShapeRouteState,
-  };
-};
-
-export const configureConfigureStepMapper = ({
-  position,
-  step,
-  updatedIntegration,
-  ...rest
-}: IEditorConfigureStep) => {
-  const { params, state } = configureIndexMapper(rest);
-  return {
-    params: {
-      ...params,
-      position,
-    } as IConfigureActionRouteParams,
-    state: {
-      ...state,
-      step,
-      updatedIntegration,
-    } as IConfigureStepRouteState,
-  };
-};
-
-export const configureTemplateStepMapper = ({
-  position,
-  step,
-  updatedIntegration,
-  ...rest
-}: IEditorConfigureStep) => {
-  const { params, state } = configureIndexMapper(rest);
-  return {
-    params: {
-      ...params,
-      position,
-    } as ITemplateStepRouteParams,
-    state: {
-      ...state,
-      step,
-      updatedIntegration,
-    } as ITemplateStepRouteState,
-  };
-};
-
-export const configureConfigureDataMapperMapper = ({
-  position,
-  step,
-  updatedIntegration,
-  ...rest
-}: IEditorConfigureStep) => {
-  const { params, state } = configureIndexMapper(rest);
-  return {
-    params: {
-      ...params,
-      position,
-    } as IDataMapperRouteParams,
-    state: {
-      ...state,
-      step,
-      updatedIntegration,
-    } as IDataMapperRouteState,
-  };
-};
-
-export function makeEditorStepRoutesResolvers(
-  esr: typeof stepRoutes
-): RouteResolver<typeof stepRoutes> {
-  return {
-    selectStep: makeResolver<
-      IEditorSelectConnection,
-      ISelectConnectionRouteParams,
-      ISelectConnectionRouteState
-    >(esr.selectStep, configureSelectConnectionMapper),
-    connection: {
-      selectAction: makeResolver<
-        IEditorSelectAction,
-        ISelectActionRouteParams,
-        ISelectConnectionRouteState
-      >(esr.connection.selectAction, configureSelectActionMapper),
-      configureAction: makeResolver<
-        IEditorConfigureAction,
-        IConfigureActionRouteParams,
-        IConfigureActionRouteState
-      >(esr.connection.configureAction, configureConfigureActionMapper),
-      describeData: makeResolver<
-        IEditorConfigureStep,
-        IDescribeDataShapeRouteParams,
-        IDescribeDataShapeRouteState
-      >(esr.connection.configureAction, configureDescribeDataShapeMapper),
-    },
-    apiProvider: {
-      upload: makeResolverNoParams(esr.apiProvider.upload),
-      review: () => 'review',
-      edit: () => 'edit',
-    },
-    basicFilter: () => 'basicFilter',
-    dataMapper: makeResolver<
-      IEditorConfigureStep,
-      IDataMapperRouteParams,
-      IDataMapperRouteState
-    >(esr.dataMapper, configureConfigureDataMapperMapper),
-    template: makeResolver<
-      IEditorConfigureStep,
-      ITemplateStepRouteParams,
-      ITemplateStepRouteState
-    >(esr.template, configureTemplateStepMapper),
-    step: makeResolver<
-      IEditorConfigureStep,
-      IConfigureStepRouteParams,
-      IConfigureStepRouteState
-    >(esr.step, configureConfigureStepMapper),
-    extension: () => 'extension',
-  };
-}
 
 // TODO: unit test every single one of these resolvers 😫
 
@@ -314,11 +80,7 @@ export const metricsResolver = makeResolver<
   },
 }));
 
-export type RouteResolver<T> = {
-  [K in keyof T]: T[K] extends string ? any : RouteResolver<T[K]>
-};
-
-const resolvers: RouteResolver<typeof routes> = {
+const resolvers = {
   list: listResolver,
   manageCicd: {
     root: manageCicdResolver,
@@ -326,7 +88,7 @@ const resolvers: RouteResolver<typeof routes> = {
   create: {
     root: makeResolverNoParams(routes.create.root),
     start: {
-      ...makeEditorStepRoutesResolvers(routes.create.start),
+      ...makeEditorResolvers(routes.create.start),
       selectStep: makeResolverNoParamsWithDefaults<
         ISelectConnectionRouteParams,
         ISelectConnectionRouteState
@@ -343,15 +105,15 @@ const resolvers: RouteResolver<typeof routes> = {
         };
       }),
     },
-    finish: makeEditorStepRoutesResolvers(routes.create.finish),
+    finish: makeEditorResolvers(routes.create.finish),
     configure: {
       root: makeResolverNoParams(routes.create.configure.root),
       index: makeResolver<IEditorIndex, IBaseRouteParams, IBaseRouteState>(
         routes.create.configure.index,
         configureIndexMapper
       ),
-      addStep: makeEditorStepRoutesResolvers(routes.create.configure.addStep),
-      editStep: makeEditorStepRoutesResolvers(routes.create.configure.editStep),
+      addStep: makeEditorResolvers(routes.create.configure.addStep),
+      editStep: makeEditorResolvers(routes.create.configure.editStep),
       saveAndPublish: makeResolver<
         IEditorIndex,
         ISaveIntegrationRouteParams,
@@ -369,8 +131,8 @@ const resolvers: RouteResolver<typeof routes> = {
         routes.integration.edit.index,
         configureIndexMapper
       ),
-      addStep: makeEditorStepRoutesResolvers(routes.integration.edit.addStep),
-      editStep: makeEditorStepRoutesResolvers(routes.integration.edit.editStep),
+      addStep: makeEditorResolvers(routes.integration.edit.addStep),
+      editStep: makeEditorResolvers(routes.integration.edit.editStep),
       saveAndPublish: makeResolver<
         IEditorIndex,
         ISaveIntegrationRouteParams,
